mod_webmail: Mark message seen based on whether frontend tells us to.
Previously, we have marked messages seen only if we were not requesting
the raw message body, since for downloading a message, we should not
change this flag. However, if actually viewing a message as raw, it
may make sense to mark it seen. There is no way for the backend to
differentiate these two cases currently. The frontend thus now tells
us explicitly whether to mark the message seen, and we use this to
mark the message seen or unseen, which makes the behavior more logical.