pty.c: Close file descriptors in off-nominal branch of PTY master thread.
All paths in the PTY master thread close both file descriptors before
breaking, except one, which can result in a file descriptor leak of
one of the file descriptors when that branch is taken. This resulted
in a leak of 1 file descriptor per remote console open at shutdown.
While that case is not a meaningful leak, we now always close both
descriptors in all branches to prevent this leak from happening.

Resolves: #20