mod_discord: Fix relaying of CTCP ACTION messages.
Fix broken parsing/generation of relayed CTCP ACTION messages from IRC.
This likely worked at one point but broken when the IRC client/relay
modules were overhauled. Now, the output is as expected again.

Also fix a bug when multiple guilds are associated with the bot token
in use, where after the first guild is processed, all remaining channel
mappings are discarded. This is because the check to verify all mappings
was done in the per-guild logic, but should only be done after all
guilds have been processed, so this logic has been moved.

In the case that there really are no mappings, we also unload the module
at this point, since there would never be anything relayed so there is
no point in continuing to run.

Resolves: #31
LBBS-42 #close