install_prereq.sh: Differentiate between Debian and Ubuntu.
The Debian-based distro version check assumed it was Debian,
which would in a shell error something like this at the
beginning of any builds on Ubuntu:

./scripts/install_prereq.sh: 179: [: Illegal number: 22.04

The direct cause of the error is because Ubuntu releases are
decimal instead of purely integer. However, the versioning
scheme is different, so this check is only appropriate for
Debian and is now targeted as such.