lock.c: Use robust mutexes when possible.
For dynamically initialized mutexes, make them robust, so that
if the owning thread were to die while holding a mutex, we have
better visibility that this happened. This is intended mainly
for ease of debugging fallout from thread cancellations, and
isn't intended to allow for graceful recovery from such incidents.