mod_smtp_delivery_external: Ignore local mailboxes in exists handler.
Explicitly ignore handling mail where tolocal is true (mail is
destined for a local mailbox). Previously, this was not done, and
for mail received from another server, mod_smtp_delivery_external
would return 1 if the sending server was authorized to receive mail.
However, this should only be done for egressing mail, not for
ingressing mail (in particular, if it is to a local mailbox, it is
not egressing). The deliver handler correctly ignored this, but the
exists handler did not, which resulted in a scenario where RCPT TO
would return success but DATA would fail with a temporary failure
code, even though a permanent failure could should have been returned
immediately.