net_imap: Improve robustness of remote IMAP client recreation.
Proxied IMAP clients are not always recreated and restored to the same
state that they were in originally. This manifests itself as receiving
"BAD Command received in Invalid state" or potentially "Must select
a mailbox first" when attempting to FETCH a message on a remote IMAP
server where the currently selected mailbox has been idle for some time.

To fix this:
* Restore the selected folder of the foreground client, so whatever
  mailbox was selected originally is reselected when recreated.
  The SELECT happens in the background, transparently.
* Fix detection of idling client when attempting to send command.
  This should never be okay so always trigger a soft assertion.