mod_webmail: Prioritize RECENT instead of EXISTS when idling.
Previously, we were mainly looking at untagged EXISTS responses
during IDLE to determine whether to send a new message notification/
preview, but this was not ideal as it allowed for many false positives.
We now use untagged RECENT responses as our main indicator. This is
more reliable, since whenever there is a new message in a folder,
the untagged EXISTS is always accompanied by an untagged RECENT anyways.

* This avoids scenarios in which an untagged EXISTS is sent by the
  server, even if there are no new messages in the mailbox. For example,
  Microsoft servers do this after an untagged EXPUNGE is sent, when a
  message is moved out of the folder to a different one. Previously,
  we would send a false positive notification for the latest message
  in the original folder when this happened.
* When we send a RECENT payload to the frontend, include the message's
  flags. This allows the frontend to know if the message is seen or not,
  and increment (or not) the folder's unseen count accordingly.