io_tls: Defer rebuilds of TLS session list.
When a TLS session ends, the entire TLS session list needs
to rebuilt so as not to exclude the session that just ended.
However, in the case of many sessions ending simultaneously,
this could be inefficient as the list would be repeatedly
for each destroyed session only to be immediately rebuilt
following the next termination. Instead, when a session
is destroyed, wait a brief amount of time for any additional
sessions to terminate, and keep waiting as long as sessions
are actively being destroyed, up to a point, at which time
the session list is rebuilt regardless. This ensures that
"starvation" of remaining active sessions cannot occur due to
them not being serviced, while eliminating most unnecessary
consecutive rebuilds of the session list.