pty.c: Open PTY slave with O_NOCTTY.
Up until now, the BBS has received a SIGHUP whenever remote consoles
exited, which originally caused the BBS to exit when the BBS was
daemonized. Workarounds were added for this, beginning with commit
5952cfbc57321101ca95c2d4c39002383bb5431d.

However, the underlying cause of the SIGHUP was a single open()
call lacking O_NOCTTY, resulting in acquisition of a controlling
terminal. By providing this flag, we prevent this from happening,
ensuring that we don't receive unwanted SIGHUP signals. Previously,
if running under systemd, because we don't daemonize, SIGHUP wasn't
ignored, and thus a remote console exiting would result in SIGHUP
killing the BBS.

We still continue to ignore SIGHUP, because when running
under systemd, the BBS still receives a SIGHUP, in the same manner
as addressed in commit ea38bdc74322a484c93b7db85eb3eb1360c7ea67.
We use SIGUSR2 for processing reloads from systemd, so there's not
much downside to ignoring SIGHUP to ensure these things aren't
conflated. However, because we don't daemonize when running under
systemd, but because STDOUT isn't a TTY in this scenario, we now also
ignore SIGHUP if STDOUT is not a TTY. In other words, it's less
whether we're daemonized and more whether STDOUT is connected to
a terminal that determines whether we ignore SIGHUP.