mod_smtp_delivery_external: Don't warn if no STARTTLS for internal hostname.
mod_smtp_delivery_external logs a warning if plaintext SMTP delivery is
attempted in most cases, except for when delivering to a private or
loopback IP address (since these are assumed to be within the trust
boundary, and encryption may not be necessary). However, we previously
logged warnings when delivering to a hostname that resolved to an internal
IP address, which if compared directly, would not have resulted in such
a warning. We now take care to resolve the hostname before checking if
it's public or not.