mod_uuid: Move uuid.c out of core into separate module.
Move uuid out of the core and into a separate module. This library was
only used for one function in the core, which was only used by one
caller in net_nntp. Support for this library is not universal (e.g.
lacking on BSD), so moving this out of the core allows the core to
build successfully on more platforms.