mod_mail: Tolerate slower responses to LOGIN/AUTHENTICATE.
Previously, we only waited 2500 ms for a successful response
to the LOGIN / AUTHENTICATE commands when logging into a
remote IMAP server. This timeout occasionally expires with
no response, particularly when talking to Gimap (Gmail).
The old timeout was unreasonably low, so increase the timeout
to 10 seconds. If we haven't gotten a response by then, then
it's likelier that something is actually wrong.

LBBS-94 #close