io_tls: Fix file descriptor leak on cleanup.
Commit 0ea87e331cfab83b8a263093b21b3da7fe734b8d introduced a regression
which leaked file descriptors for TLS sessions by adding a call to the
MARK_DEAD macro. This macro modifies the sfd variable in place, which
was not a problem in the TLS I/O thread, but would result in sfd being
NULL post-traversal in ssl_unregister_fd, causing the TLS session to
not be cleaned up. So:

* Use a local variable for the traversal in MARK_DEAD to avoid mutating
  a variable outside the macro.
* Log a warning if unregistering fails and we return -1. We were returning
  -1 previously, but there was no warning or error logged when this happened.
* Simplify some cleanup code by using PIPE_CLOSE where possible.