cli.c: Don't use errno to indicate invalid command.
Occasionally, mod_sysop will report a command was invalid,
even when it was a valid command. This can happen if errno
was last set to ENOENT, as part of whatever the CLI handler
did. Since we don't want to have to reset errno to 0 after
the CLI command returns, avoid setting errno at all and return
these error codes directly if needed to indicate a special
condition. This should eliminate "false positives" caused by
errno happening to be set to one of the special error codes.