net_ssh: Properly canonicalize nonexistent paths in SFTP.
When a REALPATH command in SFTP was processed, we were
using realpath(3) to canonicalize it, but realpath(3) only works
on paths that exist. When clients issued an SFTP REALPATH
command for a nonexistent path (such as files they were about
to upload), realpath would fail, causing the client to abort
the transfer. For paths that don't exist, we try falling
back to realpath(1), and if this fails (which it will,
if it isn't the GNU version of realpath, since only the
GNU version contains the needed option), we finally fall
back to our own custom logic to canonicalize paths.