mod_mimeparse: Don't close file descriptor already closed by GMime.
New warnings about closing file descriptors that were already closed
alerted to a file descriptor we were trying to close after the library
had already closed it for us. Since that isn't captured in the array
in fd.c, add an API to indicate a file descriptor has already been
closed outside of the BBS (e.g. in a library), so we can keep our state
up to date while avoiding double closes.