net_ftp: Fix truncated transfers when using TLS.
In the DATA_DONE_FD macro, we were closing the data connection's
actual file descriptor (datafd) before waiting for all pending
I/O data to be written (e.g. TLS encryption). This would cause
the end of the transfer (and potentially the whole transfer, for
small ones) to be lost, since the file descriptor was closed while
it was still being used. These operations are now in the right order.

This issue could be detected by checking in ssl_close that the
SSL session's file descriptor was still valid, and we now do
that to catch similar bugs.

A test capturing this issue has been added.