io.c: Add function to synchronously drain a connection.
Add a function that can wait for all I/O transformers to finish
any pending writes before returning, which can be useful to wait
synchronously for I/O to finish.

This API was initially developed as part of debugging for commit
f60618c5db5d37b125fa1449c0cb30f64f4e3c65, but is not actually needed
for that. Nonetheless, it may be useful in the future for certain
purposes, so I'm committing it anyways for potential future usage.
The changes to the bbs_io_transformation structure also give us
access to the file descriptors on both side of a transformation,
which could be useful for other purposes in the future.