io: Don't prematurely close all pipe file descriptors.
The I/O callbacks were all previously closing both sides of the read
and write pipes when the I/O thread exited. This is incorrect, as
the consumers on the other ends of the pipe may still be reading data
still in the pipe from it at the time. In the case of io_tls, this
caused intermittent test failures of test_ftps, since not all data
would get read by the FTP server before io_tls closed the read side
of that connection (when it should have only closed its write side).

To address this, we now only close the file descriptors that belong to
us in the I/O modules. The relevant I/O transformers now take
responsibility for closing the other ends, after they are actually
finished with them.

Related fixes and enhancements:
* test.c: Set a timeout for test execution (similar to the one for
  shutdown), so that if a test hangs while running, we can get
  a backtrace and exit in due time, rather than hanging indefinitely.
* thread.c: Use usleep instead of bbs_safe_sleep, since the latter
  will busy wait if called in a loop during shutdown.
* io_tls: Ensure bidirectional shutdown doesn't hang forever
  by switching the socket to nonblocking I/O.