net_ftp: Add RFC 3659 SIZE extension.
Add support for the SIZE extension to the FTP server. This is also
used in the tests to confirm an entire file has been uploaded,
before attempting to download it. This would have been useful in
recent TLS debugging, because it seemed like the download was failing,
when in fact it was not being uploaded properly in the first place.
A SIZE check prior to the RETR attempt would have made this more
obvious.