io: Allow I/O relay thread logic to be handled by io.c.
Up until now, all the I/O modules have been manually launching their
own thread to poll and manage I/O reads and writes. The structure of
this thread is almost identical between the modules, save for the
actual file descriptors and custom "I/O logic" to make the needed
transformations when reading and writing. To reduce duplicated code
and simplify the logic in I/O modules, move the module-specific stuff
into a set of callbacks so that the thread can be set up and managed
in io.c. These callbacks are optional, so it is still possible to do
things the old way if more control is needed, but the new approach is
adopted for all currently existing modules.