mod_sysop: Add tests for sysop console and terminate PTYs more cleanly.
Add some basic tests for the sysop console. This is less to
test its functionality and make sure that it works, than to
ensure that it (and the associated PTY and file descriptors)
clean up properly, since these historically have been problems.

Various cleanup has been done in mod_sysop and elsewhere to
ensure that, more than just passing, no errors or warnings
are emitted during the test indicating other minor bugs in the code.
In particular, a lot of code has been modified to ensure file
descriptors are only closed exactly once and that the test
exits quickly and reliably.

Related changes:

* node.c: Interrupt node at shutdown if shutting down asynchronously,
  to avoid blocked I/O (e.g. poll) causing shutdown to hang.
* pty.c: Don't close file descriptors that are closed elsewhere,
  and end more cleanly.
* socket.c: Fix node->skipjoin not being set in one place,
  leading to attempt to join a detached thread at node shutdown.
* thread.c: Fix incorrectly implemented timespec manipulation
  for pthread_timedjoin_np.
* mod_sysop: Join PTY thread instead of creating it detached, to
  ensure it exits before BBS shutdown finishes.
* mod_sysop: Tolerate remote consoles not having a TTY, which
  is useful for test_sysop.