node.c: Make node printout more compact.
Since not all nodes have menu/location, but a lot of space
was allocated to printing this, move this to the end of the
line and make it not fixed width anymore. In many cases,
depending on terminal width, this is enough to make each
node listing take up just one line instead of two.