net_imap: Abort if node disconnects in middle of FETCH response.
Previously, the cleanup label in process_fetch did not break out
of the loop, so even though we had detected a node disconnect, we
would still loop and try to write out the entire FETCH response
(which could potentially be quite long). On I/O error, we now
correctly abort immediately, as was originally intended.