bbs.c: Warn if core dump would not succeed.
If the BBS is started with the option to dump a core (-g), but we
determine that based on the current directory and the system's core
dump pattern, we would not be able to successfully write the core
file, then emit a warning to warn the user about this.

Also add a CLI command for testing purposes that can be used to
manually trigger a SIGABRT to test whether a core can be dumped
successfully.