nets: Return protocol error if Explicit TLS is not possible.
Some network protocol modules were not consistent in the checks
they did prior to initiating STARTTLS at a client's request.
Be consistent in always checking that the node is not already
using TLS and that TLS is available prior to honoring a STARTTLS
request. This ensures a protocol error is returned to the client,
rather than having to forcibly disconnect when STARTLS inevitably
fails in these cases.