net_smtp: Enforce different size limits for local/nonlocal delivery.
Previously, the local message size limit was enforced for all processed
messages, even submissions or relayed messages, which is inappropriate
since the recipient's server should be the only one rejecting messages
for being too large as per its local policy. To fix this, distinguish
between the max allowed size for messages delivered to local recipients
and the max allowed size for any message processed in general, and use
the right one as appropriate.