mod_smtp_delivery_external: Remove dead synchronous queuing code.
Always queue external mail for delivery and send it asynchronously.
i.e. alwaysqueue is always assumed to be true, as is sendasync,
and these settings along with mailqueue have now been removed.
This also eliminates !defined(BUGGY_SEND_IMMEDIATE) code that was
defined out.

This allows simplifying the code significantly to remove
an alternately defined version that did not work originally
and would not be very useful even if available.