mod_smtp_delivery_external: Use separate control file for queued messages.
Rather than stuffing the sender and recipient at the beginning of the
queue data file, use a separate control file, which is simpler in many
ways and more flexible, as it will allow us to store much more and much
richer information. This will give us better control over the behavior
of messages in queue and allow us to implement other extensions properly.

This change is not backwards compatible with previous queue files, so
any upgrades should be done when the queue is empty.