net_smtp: Do not dispatch DSN notifications for non-submissions.
Previously, it was possible for DSN notifications to be sent to
external senders if a delay was encountered during delivery,
even if the sender had not used the DSN extension to request such
a notification.

Modifications have also been made to store information from an SMTP
session, such as the HELO hostname and type of session, in the queue
file, so this information can be queried when generating a DSN report.