mod_discord: Add support for using webhooks to post messages.
Allow a webhook ID/token to be configured for each associated Discord
channel. This allows us to post messages and "spoof" the username,
resulting in friendlier formatted messages on the Discord side, since
they don't need to include the bot's username. For any message with a
sender on the IRC side, we can now use a webhook to post the message.

Resolves: #51