net_ssh: Don't track file descriptors used by libssh.
Commit 2ad99436d24c22316479ced54fbf8a57297f78f7 made it so that
we kept track of file descriptors opened as part of an SSH/SFTP
session. However, because libssh closes the file descriptor passed
to it, even if it did not open it, we can't reliably detect when
the file descriptor was closed and this can cause spurious test
failures. Give this up as a lost cause and disable this behavior
by default.