mod_smtp_filter: Use From address for Authenticated sender.
Use the address in the From header instead of the envelope sender
(MAIL FROM) for the Authenticated sender tag that is included in
the Received header for submissions. This is more accurate, since
the MAIL FROM for submissions has no real meaning anyways, and this
also prevents leaking information, as in the case of sending mail
using an alias for an account, the account may be used for the MAIL
FROM and disclose the primary email address anyways.