net_smtp: Relax strictness of domain verification for relaying hosts.
If domains are specified for a host in [authorized_relays], we would
attempt to deliver mail for those domains. If this list included a
hostname that was used only for HELO/EHLO but was not used publicly,
this could result in attempts to send outgoing mail for domains that
don't exist DNS, likely leading to rejections or quarantine. It is
desirable to avoid this outcome, so such hostnames should be removed
from the [authorized_relays], with only domains for which we want to
deliver mail included.

The problem with this was that the HELO hostname was being verified
strictly, using smtp_relay_authorized(). In reality, the HELO hostname
doesn't really matter too much, and if it's not a public email domain,
failure to do a strict match is expected. We therefore tolerate a loose
match, using smtp_relay_authorized_any().

This begged the question of what verification, if any, was being done
for relayed messages. In fact, a strict verification using
smtp_relay_authorized() was being done after MAIL FROM. However,
if forwarding messages without using Sender Rewriting Scheme, this
may cause rejection, since the envelope sender of the forwarded message
is left intact by the forwarding host. This is really a separate matter
from the previous two, but the strictness of this check is now
configurable, in that in the more relaxed mode, messages can be accepted
for relay from a host regardless of the envelope sender domain.