net_imap: Fix partial fetch not being applied in some cases.
Responses would include a length as if a partial response would be
returned, but then the entire chunk was sent anyways. Use that
length to truncate the response appropriately.

This was not captured in existing test cases because supersets
of test substrings will always pass, since the CLIENT_EXPECT
macros look for a substring match. As such, there isn't a good
way currently to fail a check because we got more data than we
expected. This bug thus isn't currently captured in a test case,
but it has been fixed.