net_smtp: Actually prepend headers in the correct order.
Previously, new headers would be written to a message file before
appending the existing message to it, but since the headers were
being written directly to the file, older headers would appear
before newer headers, which is wrong.

To fix this, rather than writing directly to a file, we instead
buffer the prepended headers using a stringlist, which allows us
to use a LIFO technique to prepend the headers in the correct
order.

Additionally, the Received header is now run with COMBINED scope,
rather than INDIVIDUAL scope. This ensures that it is the first
header to be prepended. This is important as logically, the header
should be first, and doesn't depend on the recipient, but also
because other COMBINED filters may use it - for example, SpamAssassin
requires our Received header to be present, or it may wrongly process
a previously existing Received header instead.

A basic test has been added that now looks for the right header order.