net_imap: Abort if node is shut down during IDLE poll.
node_shutdown interrupts the poll() system call during an IDLE poll, but
we were checking for EINTR and continuing the loop anyways. Unfortunately,
even with additional poll events, poll() doesn't return error/activity
when a shutdown occurs. Now, check if the node is inactive and abort if
so, since this indicates a shutdown was attempted on this node.