test_menus: Emulate ANSI-capable client instead of non-ANSI client.
Previously, we were pretending we didn't support ANSI to avoid
needing to do an ANSI handshake. However, this delays the test
significantly since the ANSI autodetection timeout logic needs
to expire before we can continue with the test. At the time the
test was written, test_ansi_handshake() did not exist, but now
that it does, use that to speed up the test.

Under valgrind on the same host, this reduces test execution from
about 18 seconds to around 6 seconds.