net_imap: Include RECENT in STATUS in response to remote IDLE activity.
Include RECENT in the list of attributes requested about a mailbox that
had activity as part of the information proxied to our client as part of
a NOTIFY watch. This information is then propagated all the way to any
frontend clients, and in the case of wssmail, suppresses erroneous new
message notifications when existing messages are simply marked as unread.