net_imap: Fix improper detection of currently selected folder.
In imap_notify_applicable_fetchargs, we use two different methods of
determining if an update for a folder is applicable to an IMAP session.
In the case of applicable updates, we need to know if the message
corresponds to the currently selected mailbox or some other mailbox,
e.g. to determine if we send an untagged EXISTS or untagged STATUS.

The existing logic was incomplete in that the second case handling
remote (proxied) mailboxes would compare maildir paths, which would
match if the last selected non-remote folder happened to be that path,
since this buffer isn't cleared out.

To prevent this, we now exclude remote clients from this check, to
ensure this only matches for local mailboxes in the Other Users or
Shared Folders namespaces. A test capturing this bug has also been added.