socket.c: Use a separate node lock for I/O operations.
During node shutdown, we could hang waiting to acquire the node lock
at the beginning of the shutdown process if a node was blocked on
I/O, such as in timed_write. Thus, we need to be able to interrupt
the node thread even if a node is blocked on I/O. To facilitate this,
use a separate lock for potentially prolonged I/O operations, since
this is a different locking case than normal node locking operations
anyways. This allows us to grab the node lock quickly and be able to
do the interrupt.