socket.c: Run handshake callback after node thread is started.
Previously, the protocol handshake callback (used by Telnet and RLogin)
was run by the listener thread before launching the node thread.
Do this in the new thread that is launched, so the listener thread
can go back to accepting new connections immediately.

Additionally, use a higher value of backlog for listen(), which
significantly increases the speed of test_telnet. test_telnet has
been refactored to allow changing the concurrency, though this is
left at 1 for now, since it does not appear to impact throughput.