mod_ncurses: Fix array sizing issue.
Increase the size of the arrays by one, since the definition used originally
referred to the max number of items, not the size of the array (which should
be one larger than that).

Also, within mod_operator, return -2 instead of -1 to avoid infinite loop
when the node should be disconnected.