mod_mailscript: Run maildir rules with system permissions, even if inside a mailbox.
Since MailScript rules in a maildir can only be modified by the sysop,
there isn't any reason they need to be run in the container (indeed, in
certain cases, this is too restrictive, as the sysop may want to run a
system script for a particular user's mailbox). Only run the rules found
in the user's home directory in the container.