net_ssh: Set node pointer to NULL prior to ssh_channel_close.
Fix a crash due to ssh_channel_close calling the pty_resize callback,
which results in a call to bbs_node_update_winsize if the node pointer
is non-NULL. This resulted in an assertion since we would then attempt
to lock the node, but the lock had already been destroyed.

Since the node shouldn't even be used at this point, and we don't need
to resize the terminal dimensions at connection close, set the node
pointer to NULL prior to calling this function.

LBBS-104 #close