config.c: Require configs to be explicitly unlocked when no longer needed.
Previously, configs were retrieved using bbs_config_load, but never
explicitly "released" to indicate they were no longer being used. This
allows for the small but possible chance that a config could be freed
while it is being used. To avoid this, lock configs when they are
returned and require that they be explicitly unlocked. This allows us
to avoid destroying a config while it is still being used.

A CI fix for Debian 10 is also incorporated; the Debian 10 packages are
now only available through the archive site:
https://lists.debian.org/debian-devel-announce/2025/06/msg00001.html

LBBS-105 #close