socket.c: Make connect timeout configurable.
The connect() timeout has been hardcoded to 4 seconds,
which is too short for connections that need to be brought
up on demand (e.g. dial-up) and too long for the test suite.
Make this configurable, with a new default of 30 seconds.

Also add missing break in pty.c for pty -> socket handling.