mod_smtp_delivery_external: Fix queuing bugs and improve queuing behavior.
It has been observed that other activities on a system can mess up the
"modified" timestamp, which is imperative for correct queue behavior.
For example, if the timestamp is updated even though a message in queue
was not attempted, this can result in messages stuck in queue indefinitely.

Consequently, we now store the timestamps we need in the control file itself
(which is an option that did not exist originally when there was no control
file), ensuring more reliable queue operation.

Additionally, a number of additional queue options have been added:

* Add queueimmediate option and allow queueinterval to be set to 0.
  These allow automatic and periodic queuing to be completely disabled,
  so that messages are only flushed from queue on demand, which can be
  useful in intermittently disconnected environments.
* Add ability to process just a single message in the queue.
* Add ability to delete a single message in the queue.
* Implement maxage setting and increase default from 1 day to 7 days.
* Improve display of size of messages in queue.
* A test has been added for basic testing of the queue lifecycle.