system.c: Fix implementation of waitpid with timeout.
The logic wrapping around waitpid with a timeout was implemented
incorrectly; when the timer reached 0, we meant to continue in
the loop and re-enter the timeout branch to kill the child process,
but since timeout had been modified to 0, we would instead skip
the branch with the kill and go immediately to the blocking waitpid
call, thus hanging forever.

Since the timeout mechanism was used by mod_mailscript, this could
very quickly bring SMTP deliveries to a halt and deadlock all
mail transactions.

The timeout-specific logic has been pulled out into a separate
function, which we take care not to exit until the child has
either exited or been killed, ensuring that we always return
after the timeout expires.

Additionally, bbs_sigchld_poll previously never read from the
alertpipe, which would make it return immediately and busy wait
after just one SIGCHLD; this was compounded by diffms being 0,
meaning we would loop again with the same timeout. The alertpipe
is now serviced, and the calculations use the absolute difference
from the beginning of the poll, rather than the duration of the
last poll.

Finally, in mod_mailscript, the conditions for creating a temp
file and deleting it were not exactly identical, meaning we could
attempt to delete a file by an uninitialized filename. The buffer
is now initialized to make this obvious, and the conditions are
now identical.