parallel.c: Release list lock while joining task threads.
It was possible for deadlock to occur when trying to join all the
parallel task threads, since it would hold a WRLOCK on the list,
and a task thread might still need to acquire a WRLOCK on the list
in order to finish executing. To avoid holding up any tasks that
are still running, release the lock within the loop, join the
thread, and then acquire it again.

LBBS-108 #close