mod_asterisk_queues: Improve display of agent status menu.
Instead of limiting menus to 10 columns, allow the number of options
shown simultaneously in mod_ncurses to be configurable, and make
it take up as much space as possible for agent statuses to minimize
the scrolling required.

Additionally, show active agents first, followed by inactive ones,
rather than always using the same order as in the AMI response.