net_imap: Proxy remote STATUS responses back using correct case.
Because NOTIFY is only supported for proxied servers on their INBOX
folders, the response that was proxied back to the client initially
had "INBOX" hardcoded. However, this caused issues with servers whose
INBOX had a different case, e.g. "Inbox", as the client would get
"Other Users...Inbox" in the LIST response and use that for IMAP
operations, but get "Other Users...INBOX" in the untagged STATUS
response with NOTIFY enabled.

wssmail and evergreen have both been modified to be tolerant of these
case mismatches, since the INBOX is technically case-insensitive per
RFC 3501 5.1:

https://github.com/InterLinked1/wssmail/commit/e4548597fc67e6f322f6fe09afc17b456e3f2169
https://github.com/InterLinked1/evergreen/commit/9cfbc7e54ed12f9d170f76016f77e4e3d398f164

However, the most correct thing to do here is use the original case
of the mailbox (which is already in the response, and therefore easy
to pass along at this point), rather than using a hardcoded casing.
This is also more future-proof, if non-INBOX folders are ever supported
for NOTIFY on proxied connections.