module.c: Track removed modules for bbs_module_unload().
When unloading a module with dependents (as opposed to reloading it),
a use after free could occur as bbs_module_unload() was passing a NULL
stringlist so we could not keep track if a module that needed to be
unloaded already had been, leading to use after frees. We now keep track,
just like with bbs_module_reload().

LBBS-119 #close