mod_http: Fix corrupted CGI responses.
When reading back the response from a CGI program, we used bbs_readline
to read the header lines individually, prior to dumping the body of the
response back. When checking if part of the body was already in the
readline buffer, we were making two critical mistakes:

* We were using the number of bytes in the next available line, rather
  than all available data currently in the buffer.
* Because the subsequent data hasn't been shifted to the front of the
  buffer, the beginning of "all the rest" is two bytes into the buffer,
  not immediately at the beginning.

We now use the correct length and the correct buffer starting position.

Additionally, we need to use bbs_std_close when closing file descriptors
inside the child process of the CGI program, so as not to trigger
soft assertions / backtraces.

Additionally, index.cgi is now recognized as an "index" file so that
requests to a directory root will execute it if found.